Set offset to zero (0) when reading model stream.
If using a binary field - and $this->model_data is a resource,
set the offset when reading the stream to the begining (offset 0)
each time, rather than rely on the current position.

This fixes a bug where getModel() can only be called once.

Without this, subsequent calls to getModel cuase an error as
stream_get_contents returns an empty string. The stream position is
at the end. This in turn causes unserialize to return false, causing
an error in $model->fill.